test: [ANDROAPP-6802] Moving resource manager
