+++ aggregates/src/commonMain/kotlin/org/dhis2/mobile/aggregates/model/mapper/TableModelToTableModelWithTotalRow.kt
@@ -1,9 +1,9 @@
 package org.dhis2.mobile.aggregates.model.mapper
 
-import org.dhis2.mobile.aggregates.domain.ResourceManager
 import org.dhis2.mobile.aggregates.ui.inputs.CellIdGenerator.totalCellId
 import org.dhis2.mobile.aggregates.ui.inputs.CellIdGenerator.totalHeaderRowId
 import org.dhis2.mobile.aggregates.ui.inputs.CellIdGenerator.totalId
+import org.dhis2.mobile.aggregates.ui.provider.ResourceManager
 import org.hisp.dhis.mobile.ui.designsystem.component.table.model.RowHeader
 import org.hisp.dhis.mobile.ui.designsystem.component.table.model.TableCell
 import org.hisp.dhis.mobile.ui.designsystem.component.table.model.TableModel
